Updated data types to be of stdint.

+++ D3D9_Renderer/D3D9_Renderer/source/utils/Time.h
@@ -16,8 +16,8 @@ namespace renderer
             m_tEnd = std::chrono::high_resolution_clock::now(); 
             OnTimerEndCalled();
         };
-        inline long long GetDT() const { return m_dt; };
-        inline short GetFPS() const { return m_fps; }
+        inline int64_t GetDT() const { return m_dt; };
+        inline int16_t GetFPS() const { return m_fps; }
         
     private:
         void OnTimerEndCalled();
@@ -25,9 +25,9 @@ namespace renderer
         
         std::chrono::high_resolution_clock::time_point m_tStart;
         std::chrono::high_resolution_clock::time_point m_tEnd;
-        long long m_dt;
-        long long m_systemTimer;
-        short m_numFrames;
-        short m_fps;
+		int64_t m_dt;
+		int64_t m_systemTimer;
+        int16_t m_numFrames;
+        int16_t m_fps;
 	};
 }
